refactor: Break CategoryAlertEditModal into smaller, more focused files
