+++ src/components/EditionModal/index.jsx
@@ -1,3 +1,1 @@
-export { default as AccountOrGroupSection } from './AccountOrGroupSection'
-export { default as CategorySection } from './CategorySection'
-export { default as ThresholdSection } from './ThresholdSection'
+export { default, CHOOSING_TYPES } from './EditionModal'
