+++ src/ducks/settings/CategoryAlerts/CategoryAlertEditModal.jsx
@@ -1,122 +1,14 @@
-import React, { useState } from 'react'
+import React from 'react'
+import { translate } from 'cozy-ui/react'
+import EditionModal, { CHOOSING_TYPES } from 'components/EditionModal'
 import {
-  Button,
-  Modal,
-  ModalFooter,
-  ModalButtons,
-  translate
-} from 'cozy-ui/transpiled/react'
-
-import Stepper from 'components/Stepper'
-
-import { CategoryChoice } from 'ducks/categories'
-import AccountGroupChoice from 'ducks/settings/CategoryAlerts/AccountGroupChoice'
-
-import { ModalSections } from 'components/ModalSections'
-import {
-  AccountOrGroupSection,
-  CategorySection,
-  ThresholdSection
-} from 'components/EditionModal'
-
-const CHOOSING_TYPES = {
-  category: 'category',
-  accountOrGroup: 'accountOrGroup',
-  threshold: 'threshold'
-}
-
-const SectionsPerType = {
-  [CHOOSING_TYPES.category]: CategorySection,
-  [CHOOSING_TYPES.accountOrGroup]: AccountOrGroupSection,
-  [CHOOSING_TYPES.threshold]: ThresholdSection
-}
-
-const InfoSlide = ({
-  doc,
-  fieldSpecs,
-  fieldOrder,
-  fieldLabels,
-  onRequestChooseField,
-  onChangeField
-}) => {
-  return (
-    <ModalSections>
-      {fieldOrder.map(fieldName => {
-        const FieldSection = SectionsPerType[fieldSpecs[fieldName].type]
-        const fieldSpec = fieldSpecs[fieldName]
-        const fieldLabel = fieldLabels[fieldName]
-        return (
-          <FieldSection
-            key={fieldName}
-            label={fieldLabel}
-            value={fieldSpec.getValue(doc)}
-            onClick={
-              fieldSpec.immediate ? null : () => onRequestChooseField(fieldName)
-            }
-            onChange={
-              fieldSpec.immediate
-                ? ev => {
-                    onChangeField(fieldName, ev.target.value)
-                  }
-                : null
-            }
-          />
-        )
-      })}
-    </ModalSections>
-  )
-}
-
-const ChoosingSwitch = ({ choosing }) => {
-  return (
-    <>
-      {choosing.type === CHOOSING_TYPES.category ? (
-        <CategoryChoice
-          modal={false}
-          canSelectParent={true}
-          categoryId={choosing.value.id}
-          categoryIsParent={choosing.value.isParent}
-          onSelect={choosing.onSelect}
-          onCancel={choosing.onCancel}
-        />
-      ) : null}
-      {choosing.type === CHOOSING_TYPES.accountOrGroup ? (
-        <AccountGroupChoice
-          current={choosing.value}
-          onSelect={choosing.onSelect}
-          onCancel={choosing.onCancel}
-        />
-      ) : null}
-    </>
-  )
-}
-
-const getAccountOrGroupChoiceFromAlert = alert => alert.accountOrGroup
-const getCategoryChoiceFromAlert = alert => ({
-  id: alert.categoryId,
-  isParent: alert.categoryIsParent
-})
-const getMaxThresholdFromAlert = alert => alert.maxThreshold
-
-const updatedAlertFromCategoryChoice = (initialAlert, category) => ({
-  ...initialAlert,
-  categoryIsParent: !!category.isParent,
-  categoryId: category.id
-})
-
-const updatedAlertFromAccountOrGroup = (initialAlert, accountOrGroup) => ({
-  ...initialAlert,
-  accountOrGroup: accountOrGroup
-    ? {
-        _type: accountOrGroup._type,
-        _id: accountOrGroup._id
-      }
-    : null
-})
-const updatedAlertFromMaxThresholdChoice = (initialAlert, value) => ({
-  ...initialAlert,
-  maxThreshold: parseInt(value)
-})
+  getAccountOrGroupChoiceFromAlert,
+  updatedAlertFromAccountOrGroup,
+  getCategoryChoiceFromAlert,
+  updatedAlertFromCategoryChoice,
+  getMaxThresholdFromAlert,
+  updatedAlertFromMaxThresholdChoice
+} from './helpers'
 
 const fieldSpecs = {
   accountOrGroup: {
@@ -135,82 +27,6 @@ const fieldSpecs = {
     updater: updatedAlertFromMaxThresholdChoice,
     immediate: true
   }
-}
-
-const EditionModal = props => {
-  const {
-    fieldSpecs,
-    fieldLabels,
-    fieldOrder,
-    initialDoc,
-    modalTitle,
-    okButtonLabel,
-    cancelButtonLabel,
-    onEdit,
-    onDismiss
-  } = props
-  const [doc, setDoc] = useState(initialDoc)
-  const [choosing, setChoosing] = useState(null)
-
-  const handleChoosingCancel = () => {
-    setChoosing(null)
-  }
-
-  const handleChangeField = (name, val) => {
-    const updater = fieldSpecs[name].updater
-    const updatedDoc = updater(doc, val)
-    setDoc(updatedDoc)
-  }
-
-  const handleRequestChooseField = name => {
-    const fieldSpec = fieldSpecs[name]
-    setChoosing({
-      type: fieldSpec.type,
-      value: fieldSpec.getValue(doc),
-      onSelect: val => {
-        setChoosing(null)
-        handleChangeField(name, val)
-      },
-      onCancel: handleChoosingCancel
-    })
-  }
-
-  const handleConfirmEdit = () => {
-    onEdit(doc)
-  }
-
-  return (
-    <Modal title={modalTitle} mobileFullscreen={true} dismissAction={onDismiss}>
-      <Stepper
-        showPercentage={false}
-        currentIndex={choosing ? 1 : 0}
-        onBack={() => setChoosing(null)}
-      >
-        <InfoSlide
-          doc={doc}
-          fieldOrder={fieldOrder}
-          fieldSpecs={fieldSpecs}
-          fieldLabels={fieldLabels}
-          onRequestChooseField={handleRequestChooseField}
-          onChangeField={handleChangeField}
-        />
-        <div>{choosing ? <ChoosingSwitch choosing={choosing} /> : null}</div>
-      </Stepper>
-      <ModalFooter>
-        <ModalButtons>
-          <Button
-            theme={'secondary'}
-            onClick={onDismiss}
-            label={cancelButtonLabel(props, doc)}
-          />{' '}
-          <Button
-            onClick={handleConfirmEdit}
-            label={okButtonLabel(props, doc)}
-          />
-        </ModalButtons>
-      </ModalFooter>
-    </Modal>
-  )
 }
 
 /**
